+++ sqlite3/assets/sqlite3_dart_wasm.h
@@ -35,3 +35,7 @@ int dart_sqlite3_create_collation(sqlite3 *db, const char *zName,
                                              int eTextRep, int id);
 
 int dart_sqlite3_db_config_int(sqlite3 *db, int op, int arg);
+
+int dart_sqlite3changeset_apply(sqlite3 *db, int nChangeset,
+                                           void *pChangeset, void *pCtx,
+                                           int filter);