+++ sqlite3/lib/src/implementation/session.dart
@@ -196,7 +196,7 @@ final class ChangesetIteratorImplementation implements ChangesetIterator {
               (i) => raw
                   .sqlite3changeset_old(i)
                   .okOrThrowOutsideOfDatabase(bindings)
-                  .read(),
+                  ?.read(),
             )
           : null;
       final newColumns = kind != SqliteUpdateKind.delete
@@ -205,7 +205,7 @@ final class ChangesetIteratorImplementation implements ChangesetIterator {
               (i) => raw
                   .sqlite3changeset_new(i)
                   .okOrThrowOutsideOfDatabase(bindings)
-                  .read(),
+                  ?.read(),
             )
           : null;
       current = ChangesetOperation(