+++ sqlite3/lib/src/session.dart
@@ -11,6 +11,8 @@ import 'sqlite3.dart';
 /// It is used to build changesets.
 ///
 /// Equivalent to the sqlite3_session* C object.
+///
+/// {@category common}
 abstract interface class Session {
   factory Session(CommonDatabase database, {String name = 'main'}) {
     final asImpl = database as DatabaseImplementation;
@@ -53,6 +55,12 @@ abstract interface class Session {
   bool get isNotEmpty;
 }
 
+/// A patchset obtained from a byte sequence or an active [Session].
+///
+/// Unlike [Changeset]s, patchsets don't store values for deletions (apart from
+/// the primary key) and don't keep original values for `UPDATE` records.
+///
+/// {@category common}
 abstract interface class Patchset implements Iterable<ChangesetOperation> {
   /// The binary representation of this patchset or changeset.
   Uint8List get bytes;
@@ -60,32 +68,53 @@ abstract interface class Patchset implements Iterable<ChangesetOperation> {
   @override
   ChangesetIterator get iterator;
 
+  /// Creates a patch set from the underlying serialized byte representation.
   factory Patchset.fromBytes(Uint8List bytes, CommonSqlite3 bindings) {
     return PatchsetImplementation(
         bytes, (bindings as Sqlite3Implementation).bindings);
   }
 
+  /// Applies this changeset or patchset to the [database].
   void applyTo(
     CommonDatabase database, [
     ApplyChangesetOptions options = const ApplyChangesetOptions(),
   ]);
 }
 
+/// A changeset, representing changes made while a [Session] is active.
+///
+/// {@category common}
 abstract interface class Changeset implements Patchset {
+  /// Creates a changeset from the underlying serialized byte representation.
   factory Changeset.fromBytes(Uint8List bytes, CommonSqlite3 bindings) {
     return ChangesetImplementation(
         bytes, (bindings as Sqlite3Implementation).bindings);
   }
 
+  /// Inverts this changeset.
   Changeset operator -();
 }
 
+/// A recorded operation on a changeset.
+///
+/// {@category common}
 final class ChangesetOperation {
+  /// Name of the table affected by the operation.
   final String table;
+
+  /// The amount of columns in the table, which equals the [List.length] of
+  /// [oldValues] and [newValues].
   final int columnCount;
+
+  /// The type of update made.
   final SqliteUpdateKind operation;
 
+  /// If [operation] is not [SqliteUpdateKind.insert], a list of values in the
+  /// row before the operation.
   final List<Object?>? oldValues;
+
+  /// If [operation] is not [SqliteUpdateKind.delete], a list of values in the
+  /// row after the operation.
   final List<Object?>? newValues;
 
   ChangesetOperation({
@@ -102,14 +131,21 @@ final class ChangesetOperation {
   }
 }
 
+/// An [Iterator] through a [Changeset] or [Patchset].
 abstract interface class ChangesetIterator
     implements Iterator<ChangesetOperation> {
+  /// Aborts this iterator and frees internal resources.
+  ///
+  /// This is also called automatically when [moveNext] returns false, but needs
+  /// to be called manually if the iterator is not completed.
   void finalize();
 }
 
 class ApplyChangesetOptions {
-  /// Skip changes that, when targeted table name is supplied to this function, return a truthy value.
-  /// By default, all changes are attempted.
+  /// An optional filter to only apply changes on some tables.
+  ///
+  /// When set, only operators on tables for which this function returns true
+  /// are applied.
   final bool Function(String tableName)? filter;
 
   /// Determines how conflicts are handled. **Default**: `SQLITE_CHANGESET_ABORT`.